refactor: remove block number generalisation (#201)

+++ runtime/balances_transfer_all_test.go
@@ -58,7 +58,7 @@ func Test_Balances_TransferAll_Success_AllowDeath(t *testing.T) {
 	err = ext.Encode(*encoder)
 	assert.NoError(t, err)
 
-	header := gossamertypes.NewHeader(parentHash, stateRoot, extrinsicsRoot, blockNumber, gossamertypes.NewDigest())
+	header := gossamertypes.NewHeader(parentHash, stateRoot, extrinsicsRoot, uint(blockNumber), gossamertypes.NewDigest())
 	encodedHeader, err := scale.Marshal(*header)
 	assert.NoError(t, err)
 
@@ -161,7 +161,7 @@ func Test_Balances_TransferAll_Success_KeepAlive(t *testing.T) {
 	err = ext.Encode(*encoder)
 	assert.NoError(t, err)
 
-	header := gossamertypes.NewHeader(parentHash, stateRoot, extrinsicsRoot, blockNumber, gossamertypes.NewDigest())
+	header := gossamertypes.NewHeader(parentHash, stateRoot, extrinsicsRoot, uint(blockNumber), gossamertypes.NewDigest())
 	encodedHeader, err := scale.Marshal(*header)
 	assert.NoError(t, err)
